+++ flutter_google_places_sdk_platform_interface/lib/flutter_google_places_sdk_platform.dart
@@ -6,6 +6,14 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
   FlutterGooglePlacesSdkPlatform() : super(token: _token);
 
+  static const AssetImage ASSET_POWERED_BY_GOOGLE_ON_WHITE = AssetImage(
+      'assets/google/powered_by_google_on_white.png',
+      package: 'flutter_google_places_sdk_platform_interface');
+
+  static const AssetImage ASSET_POWERED_BY_GOOGLE_ON_NON_WHITE = AssetImage(
+      'assets/google/powered_by_google_on_non_white.png',
+      package: 'flutter_google_places_sdk_platform_interface');
+
   static FlutterGooglePlacesSdkPlatform _instance =
       FlutterGooglePlacesSdkMethodChannel();