+++ flutter_google_places_sdk_platform_interface/lib/src/types/locale_util.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LocaleUtil {
-  static Map<String, String> toMap(Locale locale) => {
-        'country': locale.countryCode,
-        'language': locale.languageCode,
-      };
-}