+++ flutter_google_places_sdk_platform_interface/lib/src/types/types.dart
@@ -1,3 +1,2 @@
 export 'autocomplete_predictions.dart';
 export 'find_autocomplete_predictions_response.dart';
-export 'locale_util.dart';