+++ flutter_google_places_sdk_platform_interface/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0+1
+
+* Added 2 assets (powered by google white/non-white variations)
+* Added test for the 2 assetse
+* `LocaleUtils` was removed and function as inlined where it's used (it had a single use only)
+
 ## 0.1.0
 
 * Initial Open Source release.